Address node demo review feedback
