+++ .github/workflows/qaci.yml
@@ -67,6 +67,10 @@ jobs:
         run: cargo check --target=wasm32-unknown-unknown
         working-directory: ./examples/node-demo
 
+      - name: Run clippy for unified browser node demo
+        run: cargo clippy --target=wasm32-unknown-unknown --all-targets -- -D warnings
+        working-directory: ./examples/node-demo
+
       - name: Run unified browser node demo tests
         uses: coactions/setup-xvfb@v1
         with:
