+++ .gitignore
@@ -5,6 +5,8 @@ config*.yaml
 **/*.rs.bk
 **/*.ipynb
 **/*.js
+!examples/node-demo/extension-assets/
+!examples/node-demo/extension-assets/*.js
 .ipynb_checkpoints
 # Local environment configuration
 .envrc
