+++ examples/node-demo/src/hex.rs
@@ -1,0 +1,10 @@
+//! Shared hexadecimal parsing helpers.
+
+pub(crate) fn hex_nibble(byte: u8) -> Option<u8> {
+    match byte {
+        b'0'..=b'9' => Some(byte - b'0'),
+        b'a'..=b'f' => Some(byte - b'a' + 10),
+        b'A'..=b'F' => Some(byte - b'A' + 10),
+        _ => None,
+    }
+}
