+++ examples/node-demo/src/lib.rs
@@ -12,6 +12,7 @@ mod custom;
 mod dweb;
 mod extension;
 mod forms;
+mod hex;
 mod node;
 mod peer_sync;
 mod proof;
